pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-based-ai-assistant"
version = "0.1.0"
description = "Production-ready modular RAG Assistant with FastAPI and Streamlit"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "RAG Team"}]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.5.0",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-community>=0.3.0",
  "chromadb>=0.5.0",
  "faiss-cpu>=1.8.0.post1",
  "pypdf>=5.0.0",
  "unstructured>=0.16.0",
  "langchain-text-splitters>=0.3.0",
  "streamlit>=1.40.0",
  "httpx>=0.27.0",
  "tenacity>=9.0.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.8",
  "mypy>=1.11.0"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
